fixed: cli -> password manager file based

+++ fortipass-cli/src/key_creator.rs
@@ -13,8 +13,13 @@ impl Creator for KeyCreator {
     type Manager = KeyManager;
     type Return = [u8; 32];
 
-    fn create(&self, manager: &Self::Manager, file_manager: &FileManager, filename: &str, _: Option<&str>) -> io::Result<()> {
-        let path = file_manager.secrets_path.join(filename);
+    fn create(&self, manager: &Self::Manager, file_manager: &FileManager) -> io::Result<()> {
+        let path = file_manager.secrets_path.join(file_manager.key_name);
+
+        if path.is_file() {
+            return Err(io::Error::from(io::ErrorKind::AlreadyExists))
+        }
+
         let key = manager.generate_key();
 
         let mut buf = fs::File::create(path)?;
@@ -45,15 +50,15 @@ mod tests {
 
     #[test]
     fn getting_key() {
-        let file_manager = FileManager::new(&Path::new("../.secrets"));
+        let file_manager = FileManager::new(&Path::new("../.secrets"), "key");
         let key_manager = KeyManager;
         let key_creator = KeyCreator;
 
         let mut key = [0u8; 32];
 
         if !file_manager.secrets_path.is_dir() {
-            key_creator.create(&key_manager, &file_manager, "key", None).expect("Failed creating key file.");
-            key = key_creator.retrieve(&key_manager, &file_manager, "key").expect("Faled getting key.");
+            key_creator.create(&key_manager, &file_manager).expect("Failed creating key file.");
+            key = key_creator.retrieve(&key_manager, &file_manager, file_manager.key_name).expect("Faled getting key.");
         }
 
         assert_eq!(32, key.len());
